+++ src/js/libpannellum.js
@@ -1,6 +1,6 @@
 /*
  * libpannellum - A WebGL and CSS 3D transform based Panorama Renderer
- * Copyright (c) 2012-2014 Matthew Petroff
+ * Copyright (c) 2012-2015 Matthew Petroff
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,15 +22,19 @@
  */
 
 window.libpannellum = (function(window, document, undefined) {
+
+'use strict';
+
 /* Image Type argument can be that of "equirectangular" or "cubemap".
  * If "cubemap" is used, the image argument should be an array of images
  * instead of a single image.  They should be the order of:
  * +z, +x, -z, -x, +y, -y.
  */
-function Renderer(container, image, imageType) {
+function Renderer(container, image, imageType, video) {
     this.container = container;
     this.canvas = container.querySelector('#canvas');
     this.image = image;
+    this.video = video;
 
     // Default argument for image type
     this.imageType = 'equirectangular';
@@ -41,6 +45,8 @@ function Renderer(container, image, imageType) {
     var program, gl;
 
     this.init = function(haov, vaov, voffset) {
+        var s;
+        
         // Enable WebGL on canvas
         gl = this.canvas.getContext('experimental-webgl', {alpha: false, depth: false});
         
@@ -56,7 +62,7 @@ function Renderer(container, image, imageType) {
             // Add images
             var path = this.image.basePath + this.image.fallbackPath;
             var sides = ['f', 'b', 'u', 'd', 'l', 'r'];
-            for (var s = 0; s < 6; s++) {
+            for (s = 0; s < 6; s++) {
                 this.world.querySelector('.' + sides[s] + 'face').style.backgroundImage = 'url("' + path.replace('%s',sides[s]) + '.' + this.image.extension + '")';
             }
             
@@ -65,12 +71,16 @@ function Renderer(container, image, imageType) {
             console.log('Error: no WebGL support detected!');
             throw {type: 'no webgl'};
         }
-        this.image.fullpath = this.image.basePath + this.image.path;
+        if (this.image.basePath) {
+            this.image.fullpath = this.image.basePath + this.image.path;
+        } else {
+            this.image.fullpath = this.image.path;
+        }
         this.image.invTileResolution = 1 / this.image.tileResolution;
         
         var vertices = this.createCube();
         this.vtmp = [];
-        for ( var s = 0; s < 6; s++ ) {
+        for (s = 0; s < 6; s++) {
             this.vtmp[s] = vertices.slice(s * 12, s * 12 + 12);
             vertices = this.createCube();
         }
@@ -206,17 +216,18 @@ function Renderer(container, image, imageType) {
         
         // Check if there was an error
         if (gl.getError() !== 0) {
+            var width, maxWidth;
             console.log('Error: Something went wrong with WebGL!');
             if (this.imageType == 'equirectangular') {
-                var width = Math.max(this.image.width, this.image.height);
-                var maxWidth = gl.getParameter(gl['MAX_TEXTURE_SIZE']);
+                width = Math.max(this.image.width, this.image.height);
+                maxWidth = gl.getParameter(gl.MAX_TEXTURE_SIZE);
                 if (width > maxWidth) {
                     console.log('Error: The image is too big; it\'s ' + width + 'px wide, but this device\'s maximum supported width is ' + maxWidth + 'px.');
                     throw {type: 'webgl size error', width: width, maxWidth: maxWidth};
                 }
             } else if (this.imageType == 'cubemap') {
-                var width = this.image[0].width;
-                var maxWidth = gl.getParameter(gl['MAX_CUBE_MAP_TEXTURE_SIZE']);
+                width = this.image[0].width;
+                maxWidth = gl.getParameter(gl.MAX_CUBE_MAP_TEXTURE_SIZE);
                 if (width > maxWidth) {
                     console.log('Error: The cube face image is too big; it\'s ' + width + 'px wide, but this device\'s maximum supported width is ' + maxWidth + 'px.');
                     throw {type: 'webgl size error', width: width, maxWidth: maxWidth};
@@ -227,9 +238,11 @@ function Renderer(container, image, imageType) {
     };
 
     this.render = function(pitch, yaw, hfov) {
+        var focal;
+        
         // If no WebGL
         if (!gl && this.imageType == 'multires') {
-            var focal = 1 / Math.tan(hfov / 2);
+            focal = 1 / Math.tan(hfov / 2);
             var zoom = focal * this.canvas.width / 2 + 'px';
             var transform = 'translate3d(0px, 0px, ' + zoom + ') rotateX(' + pitch + 'rad) rotateY(' + yaw + 'rad) rotateZ(0rad)';
             this.world.style.webkitTransform = transform;
@@ -242,12 +255,20 @@ function Renderer(container, image, imageType) {
         if (this.imageType != 'multires') {
             // Calculate focal length from vertical field of view
             var vfov = 2 * Math.atan(Math.tan(hfov * 0.5) / (this.canvas.width / this.canvas.height));
-            var focal = 1 / Math.tan(vfov * 0.5);
+            focal = 1 / Math.tan(vfov * 0.5);
             
             // Pass psi, theta, and focal length
             gl.uniform1f(program.psi, yaw);
             gl.uniform1f(program.theta, pitch);
             gl.uniform1f(program.f, focal);
+            
+            if (this.video === true) {
+                // Update texture if video
+                if (this.imageType == 'equirectangular') {
+                    gl.bindTexture(gl.TEXTURE_2D, program.texture);
+                    gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGB, gl.RGB, gl.UNSIGNED_BYTE, this.image);
+                }
+            }
             
             // Draw using current buffer
             gl.drawArrays(gl.TRIANGLES, 0, 6);
@@ -272,8 +293,8 @@ function Renderer(container, image, imageType) {
             // Find current nodes
             var rotPersp = this.rotatePersp(perspMatrix, matrix);
             program.nodeCache.sort(this.multiresNodeSort);
-            if (program.nodeCache.length > 200
-                && program.nodeCache.length > program.currentNodes.length + 50) {
+            if (program.nodeCache.length > 200 &&
+                program.nodeCache.length > program.currentNodes.length + 50) {
                 // Remove older nodes from cache
                 program.nodeCache.splice(200, program.nodeCache.length - 200);
             }
@@ -384,12 +405,12 @@ function Renderer(container, image, imageType) {
             
             // Add node to current nodes and load texture if needed
             var inCurrent = false;
-            for (var i = 0; i < program.nodeCache.length; i++) {
-                if (program.nodeCache[i].path == node.path) {
+            for (var k = 0; k < program.nodeCache.length; k++) {
+                if (program.nodeCache[k].path == node.path) {
                     inCurrent = true;
-                    program.nodeCache[i].timestamp = program.nodeCacheTimestamp++;
-                    program.nodeCache[i].diff = node.diff;
-                    program.currentNodes.push(program.nodeCache[i]);
+                    program.nodeCache[k].timestamp = program.nodeCacheTimestamp++;
+                    program.nodeCache[k].diff = node.diff;
+                    program.currentNodes.push(program.nodeCache[k]);
                     break;
                 }
             }
@@ -475,8 +496,8 @@ function Renderer(container, image, imageType) {
                     ntmp = new MultiresNode(vtmp, node.side, node.level + 1, node.x*2+1, node.y*2, this.image.fullpath);
                     children.push(ntmp);
                 }
-                if (!(node.x == numTiles && doubleTileSize < this.image.tileResolution)
-                    && !(node.y == numTiles && doubleTileSize < this.image.tileResolution)) {
+                if (!(node.x == numTiles && doubleTileSize < this.image.tileResolution) &&
+                    !(node.y == numTiles && doubleTileSize < this.image.tileResolution)) {
                     vtmp = [v[0]*f1+v[6]*i1,  v[1]*f2+v[7]*i2,  v[2]*f3+v[8]*i3,
                               v[3]*f+v[6]*i,  v[4]*f2+v[7]*i2,  v[5]*f3+v[8]*i3,
                                        v[6],             v[7],             v[8],
@@ -606,9 +627,8 @@ function Renderer(container, image, imageType) {
         // Find optimal level
         var newLevel = 1;
         while ( newLevel < this.image.maxLevel &&
-            this.canvas.width > this.image.cubeResolution
-            * Math.pow(2, newLevel - this.image.maxLevel)
-            * hfov / (Math.PI / 2) * 0.9 ) {
+            this.canvas.width > this.image.tileResolution *
+            Math.pow(2, newLevel - 1) * Math.tan(hfov / 2) * 0.707 ) {
             newLevel++;
         }
         
@@ -816,8 +836,8 @@ var fragMulti = [
 ].join('');
 
 return {
-    renderer: function(canvas, image, imagetype) {
-        return new Renderer(canvas, image, imagetype);
+    renderer: function(canvas, image, imagetype, video) {
+        return new Renderer(canvas, image, imagetype, video);
     }
 };