+++ changelog.md
@@ -2,6 +2,56 @@ Changelog
 =========
 
 
+Changes in Pannellum 2.1.1
+--------------------------
+
+Bugfixes:
+ - Force subpixel rendering for hot spots
+
+
+Changes in Pannellum 2.1.0
+--------------------------
+
+New Features:
+
+ - Ability to limit pitch, yaw, and hfov extents
+ - Can set starting pitch and yaw in scene linked to by hotspot
+ - Pinch to zoom
+ - Zoom and fullscreen controls can be hidden
+ - "Inertia"
+ - Option to begin auto rotating after a period of user inactivity
+ - Use Photo Sphere XMP metadata for configuration
+ - Preliminary equirectangular video support (no controls)
+
+Improvements:
+
+ - Loading progress bar is displayed for equirectangular panoramas
+ - Error message for image being to large for a device is now much more
+   descriptive
+ - Zoom level choosing for multiresolution panoramas is improved
+ - Documentation of configuration parameters was added
+ - Python 2.7 support for multiresolution tile generator script
+
+Bugfixes:
+
+ - Fix bug where preview images wasn't always loaded for cubic panoramas
+ - Hot spots are now displayed behind controls
+ - Fix bug with multiresolution panoramas when `basePath` isn't defined
+ - Error message displayed for IE 9
+
+Backwards-Incompatible Configuration Parameter Changes:
+
+ - `voffset` changed to `vOffset`
+ - `autorotate` changed to `autoRotate`
+ - `autoload` changed to `autoLoad`
+ - `autoLoad` value changed from `'yes'` to `true`
+
+Other:
+
+ - Popout mode, for browsers that do not support the fullscreen API, has been
+   removed
+
+
 Changes in Pannellum 2.0.1
 --------------------------